# Array: optimize shift and unshift