@bcardiff 2020-12-15
I believe that in the `CrystalArraySyntheticProvider`, the `self.`buffer initialization should change to `self.buffer = self.valobj.child[3]`.