@bcardiff 2020-12-15
At https://github.com/crystal-lang/crystal/blob/master/etc/lldb/crystal_formatters.py#L14

Unfortunately, the manual specs that can be trigger by https://github.com/crystal-lang/crystal/blob/master/spec/debug/test.sh do not include a spec for Array.