@asterite 2020-12-15
@bcardiff Good catch, done!

We can't change the ivars order because then the Array will occupy more space.