@bcardiff 2020-12-15
FYI, The linux_32 failure seems to be because of a different value of `__pad1 : UShort` which should not matter and has probably random date. But default struct equality compares all fields.