@RespiteSage 2020-12-15
This is exciting! However, how does `Deque` compare now to `Array`? Is there still a performance benefit, or does its usefulness become limited to a more focused API?