@mwlang 2020-12-16
This is a big deal.  I would love to see this particular optimization make it into the merge.  I am _constantly_ having to refactor code that uses Array's to find different solutions than what comes most naturally.  Perhaps I'm just spoiled by Ruby's implementation of Arrays and I need to be less lazy, but old habits die hard and one great reason to use Crystal is precisely because so many Ruby concepts carry over.