@asterite 2020-12-17
> However, how does Deque compare now to Array?

Good question. You can try doing some benchmarks on both data structures and see.

My guess is that Deque will be more memory efficient when used exclusively as a deque, compared to Array, but I don't know. For example rotate is more efficient in Deque. I wouldn't remove Deque from the standard library.